Take the Conekta public key from admin
The solidus_conekta_card.js file needs the Conekta public key from
the user's Conekta account. The user adds this key to the admin when
creating a new payment with Conetka card. Here we take the value that
the user entered from the admin and set it on the Conekta.js library.